All tunable strings in one config. Add comments, add SD support, add buttons and speed hall support rework display dashboard screen

+++ src/battery.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int battery_voltage_to_percent(float voltage);
+float battery_voltage_to_percent(float voltage);
